backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
pymysql
passlib
flask-cors
python-dotenv
gunicorn